move MockSessionManagerObserver
